backend/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
sqlalchemy==2.0.41
psycopg2-binary==2.9.9
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.3.0
python-multipart==0.0.20
python-dotenv==1.1.0
pydantic[email]==2.11.7
websockets==15.0.1
pytest==8.4.2
httpx==0.28.1
